moved download_unread_filesystem_id to journalist_app/main.py
